rename support.py to _support.py to avoid ambiguous when import
