+++ Lib/test/test_ctypes/test_struct_fields.py
@@ -1,6 +1,6 @@
 import unittest
 from ctypes import Structure, Union, sizeof, c_char, c_int
-from .support import (CField, Py_TPFLAGS_DISALLOW_INSTANTIATION,
+from ._support import (CField, Py_TPFLAGS_DISALLOW_INSTANTIATION,
                       Py_TPFLAGS_IMMUTABLETYPE)
